Mostly fixes to Psychology systems
Elections will now notify you if there is a tie, to make it clear that
it's not a bug.
Pawns hanging out together will no longer try and reserve the same item
to do the same joy activity together.
Fixed a bug that was causing elections to be won by the most competitive
candidate, rather than the one with the most votes.
Colonists will now hit on psychopaths (who never had 5+ opinion of
anyone, mostly) if they find them attractive enough.
Hopefully fixed the incompatibility between Psychology and Zombieland. I
will need more input from the mod's creator to ensure this is resolved.