Mayoral offices and several improvements
Fixed #16. Colonists should now be able to hold parties again.
Fixed #20. Romantic partners will no longer cancel their dates when they schedule them for times when they would normally be able to go on a date, but wouldn't be able to hold a party.
Closed #15. Exposed the text of the Paranoid and Fell Plotting mental breaks to translation.
Closed #21. You can now designate a room as an office by placing a table in it, and then selecting the table and activating the room as a mayoral office through it. No extra furniture necessary. Mayors will prefer to hold meetings in offices, but they'll still hold them in their bedroom if no office is available.
Fixed a mistake in the xpath patches that prevented conflicting traits from being added to certain vanilla traits.
You can now change the duration of the opinion modifier from conversations in the mod settings.