Mayoral offices fixed
Fixed #23. No more NPE on maps with no colonists in them.
Fixed an NPE upon treating a pawn's traits.
Fixed the bug that was preventing mayors from holding meetings in their offices.
Removed the hotkey for designating a room as an office because it's something you're gonna do so rarely anyway.