Improvements to colonist relations
Colonists will now get small bonuses to opinion of each other for hanging out or going on a date, regardless of conversations that happen during it (this was intended behavior for dates, but it fell through the cracks).
Pawns that cannot speak will no longer have conversations.
Colonists that are hanging out or going on a date are now much more likely to have conversations with each other.
Colonists with relationships that are lackluster are now more likely to plan dates.
Some improvements to the hanging out code.
Fixed a bug when saving the game during voting for an election.
When a meeting starts, the mayor and constituent will now interrupt their previous jobs immediately so they are more likely to attend in time.