Lord job fixes
Increased the mood bonus for Pluviophiles being outside so that it counteracts the "soaking wet" malus.
Pluviophiles should no longer get a rain mood bonus from the snow.
Fixed a typo in the mayoral candidate letter.
Restored the 2 season opinion penalty for recruiting prisoners, but made sure that it only applies to recruited prisoners, not all recruitments.
Restored the "Mayor" title to mayors.
Some fixes to dates/hanging out.
Added log messages to some causes that would make an election fail to start, for the player's sake.
Increased the strength of constituent visit thoughts based on how long they're both in the meeting.