More mayoral fixes
Fixed another zeroing-out glitch in the equations for the mood bonus a colonist gets for complaining to the mayor. Colonists who complain to the mayor will now get a proper mood bonus for it. Remember, the fancier the office, and the more Empathetic the mayor, the higher the bonus.
Fixed the Sedentary trait to update the "settled in" thought properly.
Prudes who see a naked person will now properly get the negative thought for it.
Fixed some errors when pawns try to do joy activities while hanging out.
Added a letter sent when a colonist gets anxiety from a mental break.
Fixed a glitch that could cause pawns to have deep talks instead of conversations.